Move all qbft encoding, decoding and root methods to a separate json_testutils file
